Drop support for React 15
