+++ packages/react-jsx-highcharts/package.json
@@ -105,8 +105,8 @@
   },
   "peerDependencies": {
     "highcharts": "^5.0.0 || ^6.0.0",
-    "react": "^15.0.0 || ^16.0.0",
-    "react-dom": "^15.0.0 || ^16.0.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "prop-types": "^15.0.0"
   }
 }
