+++ packages/react-jsx-highcharts/src/components/Axis/index.js
@@ -1,7 +1,7 @@
 import Axis from './Axis';
 import AxisTitle from './AxisTitle';
-import provideChart from '../ChartProvider/index2';
-import provideAxis from '../AxisProvider/index2';
+import provideChart from '../ChartProvider';
+import provideAxis from '../AxisProvider';
 const ChartAxis = provideChart(Axis);
 ChartAxis.Title = provideAxis(AxisTitle);
 export default ChartAxis;
