+++ packages/react-jsx-highcharts/src/components/BarSeries/index.js
@@ -1,3 +1,3 @@
 import BarSeries from './BarSeries';
-import provideChart from '../ChartProvider/index2';
+import provideChart from '../ChartProvider';
 export default provideChart(BarSeries);
