+++ packages/react-jsx-highcharts/src/components/Chart/index.js
@@ -1,3 +1,3 @@
 import Chart from './Chart';
-import provideChart from '../ChartProvider/index2';
+import provideChart from '../ChartProvider';
 export default provideChart(Chart);
