+++ packages/react-jsx-highcharts/src/components/Credits/index.js
@@ -1,3 +1,3 @@
 import Credits from './Credits';
-import provideChart from '../ChartProvider/index2';
+import provideChart from '../ChartProvider';
 export default provideChart(Credits);
