+++ packages/react-jsx-highcharts/src/components/Debug/index.js
@@ -1,3 +1,3 @@
 import Debug from './Debug';
-import provideChart from '../ChartProvider/index2';
+import provideChart from '../ChartProvider';
 export default provideChart(Debug);
