+++ packages/react-jsx-highcharts/src/components/Hidden/Hidden.js
@@ -11,17 +11,9 @@ class Hidden extends Component {
     const { children } = this.props;
     if (!children) return null;
 
-    if (React.version >= '16.0.0') {
-      return [
-        children
-      ];
-    }
-
-    return (
-      <div style={{ display: 'none' }}>
-        {children}
-      </div>
-    );
+    return [
+      children
+    ];
   }
 }
