+++ packages/react-jsx-highcharts/src/components/HighchartsChart/index.js
@@ -1,3 +1,3 @@
 import HighchartsChart from './HighchartsChart';
-import provideHighcharts from '../HighchartsProvider/index2';
+import provideHighcharts from '../HighchartsProvider';
 export default provideHighcharts(HighchartsChart);
