+++ packages/react-jsx-highcharts/src/components/Legend/index.js
@@ -1,6 +1,6 @@
 import Legend from './Legend';
 import LegendTitle from './LegendTitle';
-import provideChart from '../ChartProvider/index2';
+import provideChart from '../ChartProvider';
 const ChartLegend = provideChart(Legend);
 ChartLegend.Title = provideChart(LegendTitle);
 export default ChartLegend;
