+++ packages/react-jsx-highcharts/src/components/Loading/index.js
@@ -1,3 +1,3 @@
 import Loading from './Loading';
-import provideChart from '../ChartProvider/index2';
+import provideChart from '../ChartProvider';
 export default provideChart(Loading);
