+++ packages/react-jsx-highcharts/src/components/Options3d/index.js
@@ -1,3 +1,3 @@
 import Options3d from './Options3d';
-import provideChart from '../ChartProvider/index2';
+import provideChart from '../ChartProvider';
 export default provideChart(Options3d);
