+++ packages/react-jsx-highcharts/src/components/PlotBand/index.js
@@ -1,6 +1,6 @@
 import PlotBand from './PlotBand';
 import PlotBandLabel from './PlotBandLabel';
-import provideAxis from '../AxisProvider/index2';
+import provideAxis from '../AxisProvider';
 const ChartPlotBand = provideAxis(PlotBand);
 ChartPlotBand.Label = provideAxis(PlotBandLabel);
 export default ChartPlotBand;
