+++ packages/react-jsx-highcharts/src/components/PlotLine/index.js
@@ -1,6 +1,6 @@
 import PlotLine from './PlotLine';
 import PlotLineLabel from './PlotLineLabel';
-import provideAxis from '../AxisProvider/index2';
+import provideAxis from '../AxisProvider';
 const ChartPlotLine = provideAxis(PlotLine);
 ChartPlotLine.Label = provideAxis(PlotLineLabel);
 export default ChartPlotLine;
