+++ packages/react-jsx-highcharts/src/components/Series/Series.js
@@ -1,4 +1,4 @@
-import { Component } from 'react';
+import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import isEqual from 'lodash/isEqual';
 import isImmutable from 'is-immutable';
