+++ packages/react-jsx-highcharts/src/components/Series/index.js
@@ -1,3 +1,3 @@
 import Series from './Series';
-import provideAxis from '../AxisProvider/index2';
+import provideAxis from '../AxisProvider';
 export default provideAxis(Series);
