+++ packages/react-jsx-highcharts/src/components/Subtitle/index.js
@@ -1,3 +1,3 @@
 import Subtitle from './Subtitle';
-import provideChart from '../ChartProvider/index2';
+import provideChart from '../ChartProvider';
 export default provideChart(Subtitle);
