+++ packages/react-jsx-highcharts/src/components/Tooltip/index.js
@@ -1,3 +1,3 @@
 import Tooltip from './Tooltip';
-import provideChart from '../ChartProvider/index2';
+import provideChart from '../ChartProvider';
 export default provideChart(Tooltip);
