+++ packages/react-jsx-highcharts/src/components/XAxis/index.js
@@ -1,5 +1,5 @@
 import XAxis from './XAxis';
-import provideChart from '../ChartProvider/index2';
+import provideChart from '../ChartProvider';
 import Axis from '../Axis';
 const ChartXAxis = provideChart(XAxis);
 ChartXAxis.Title = Axis.Title;
